update(typing): move interface
